Move dense-specific methods out of Coordinate base class.
